code refactoring, CSimReads read simulation class now part of libkit4b
