+++ blast2csv/blast2csv.vcxproj
@@ -168,6 +168,7 @@
       <PrecompiledHeader>Use</PrecompiledHeader>
       <WarningLevel>Level3</WarningLevel>
       <DebugInformationFormat>ProgramDatabase</DebugInformationFormat>
+      <FunctionLevelLinking>true</FunctionLevelLinking>
     </ClCompile>
     <Link>
       <AdditionalDependencies>shlwapi.lib;..\libkit4b\release64\libkit4b.lib;..\libzlib\release64\libzlib.lib;%(AdditionalDependencies)</AdditionalDependencies>
