+++ libkit4b/kit4bversion.h
@@ -1,3 +1,3 @@
 #pragma once
 // this file is updated each time a release is made
-const char kit4bversion[] = "0.4.6";
+const char kit4bversion[] = "0.4.8";
