+++ libkit4b/libkit4b.vcxproj
@@ -223,6 +223,7 @@
     <ClInclude Include="Shuffle.h" />
     <ClInclude Include="SimpleGlob.h" />
     <ClInclude Include="SimpleRNG.h" />
+    <ClInclude Include="SimReads.h" />
     <ClInclude Include="SmithWaterman.h" />
     <ClInclude Include="sqlite3.h" />
     <ClInclude Include="Stats.h" />
@@ -278,6 +279,7 @@
     <ClCompile Include="SfxArray.cpp" />
     <ClCompile Include="Shuffle.cpp" />
     <ClCompile Include="SimpleRNG.cpp" />
+    <ClCompile Include="SimReads.cpp" />
     <ClCompile Include="SmithWaterman.cpp" />
     <ClCompile Include="sqlite3.c">
       <CompileAs Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">CompileAsC</CompileAs>
