+++ locmarkers/LocKMers.h
@@ -9,8 +9,6 @@ const int cMinHamming = 1;				// minimum required Hamming separation
 const int cDfltHamming = 2;				// default Hamming separation
 const int cMaxHamming = 5;				// maximum required Hamming separation
 
-const int cMaxWorkerThreads = 128;			// limiting max number of threads to this many
-
 const int cBlockReqSize = 0x07ffff;		// each worker thread will request a block containing no more than this total concatenated sequences length
 const int cConcatSeqBuffSize = (cBlockReqSize * cMaxWorkerThreads * 4);	// will buffer up to this sized concatenated sequences from which blocks will be allocated
